+++ packages/libs/sdk/src/api/graphql/queries/fragments/EventsByNft.ts
@@ -1,6 +1,6 @@
 import { gql } from '@apollo/client/core';
 
-import { TokenEventInfo } from './tokenEvent';
+import { TokenEventInfo } from './nodes/tokenEvent';
 import { Pagination } from './pagination';
 
 export const NftEventsFragment = gql`
