+++ packages/libs/sdk/src/api/graphql/queries/fragments/getBalancesByWalletENS.ts
@@ -1,6 +1,6 @@
 import { gql } from '@apollo/client/core';
 import { Pagination } from './pagination';
-import { TokenBalanceNode } from './tokenBalanceNode';
+import { TokenBalanceNode } from './nodes/tokenBalanceNode';
 
 export const GetBalancesByWalletENSFragment = gql`
   fragment GetBalancesByWalletENSFragment on EVMSchemaType {
