+++ packages/libs/sdk/src/api/graphql/queries/fragments/nftsByWalletAddress.ts
@@ -1,6 +1,6 @@
 import { gql } from '@apollo/client/core';
 
-import { WalletNFTNode } from './WalletNft';
+import { WalletNFTNode } from './nodes/WalletNft';
 import { Pagination } from './pagination';
 
 export const WalletByAddressFragment = gql`
