+++ packages/libs/sdk/src/api/graphql/queries/fragments/nodes/ERC1155Node.ts
