+++ packages/libs/sdk/src/api/graphql/queries/fragments/nodes/ERC721Node.ts
