+++ packages/libs/sdk/src/api/graphql/queries/fragments/nodes/TrendingCollection.ts
